tec: Split Links controller file
