+++ src/Application.php
@@ -78,10 +78,12 @@ class Application
         $router->addRoute('post', '/links/:id/edit', 'Links#update', 'update link');
         $router->addRoute('get', '/links/:id/fetch', 'Links#showFetch', 'show fetch link');
         $router->addRoute('post', '/links/:id/fetch', 'Links#fetch', 'fetch link');
-        $router->addRoute('get', '/links/:id/collections', 'Links#collections', 'link collections');
-        $router->addRoute('post', '/links/:id/collections', 'Links#updateCollections', 'update link collections');
-        $router->addRoute('post', '/links/:id/bookmark', 'Links#bookmark', 'bookmark link');
-        $router->addRoute('post', '/links/:id/unbookmark', 'Links#unbookmark', 'unbookmark link');
+
+        // Link collections
+        $router->addRoute('get', '/links/:id/collections', 'LinkCollections#index', 'link collections');
+        $router->addRoute('post', '/links/:id/collections', 'LinkCollections#update', 'update link collections');
+        $router->addRoute('post', '/links/:id/bookmark', 'LinkCollections#bookmark', 'bookmark link');
+        $router->addRoute('post', '/links/:id/unbookmark', 'LinkCollections#unbookmark', 'unbookmark link');
 
         // Messages
         $router->addRoute('get', '/links/:link_id/messages', 'LinkMessages#index', 'links/messages');
